# specs/014-bot-plugin-platform/contracts/plugin-runtime-api.openapi.yaml
openapi: 3.0.3
info:
  title: Korus Plugin Runtime API
  version: 1.0.0
  description: Spec 014 — language-agnostic sidecar/bridge contract (integrations VM)
paths:
  /v1/plugin/handle:
    post:
      summary: Handle plugin event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginEvent'
      responses:
        '200':
          description: Plugin response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginResponse'
  /health:
    get:
      summary: Health probe
      responses:
        '200':
          description: OK
components:
  schemas:
    PluginEvent:
      type: object
      required: [event_id, type]
      properties:
        event_id: { type: string }
        bot_instance_id: { type: string, format: uuid }
        plugin_class: { type: string, enum: [L0, L1, L2, L3] }
        type: { type: string, enum: [mention, slash, button, schedule, external_webhook] }
        user_id: { type: string, format: uuid }
        chat_id: { type: string, format: uuid }
        text: { type: string }
        payload: { type: object, additionalProperties: true }
        config_snapshot: { type: object, additionalProperties: true }
    PluginMessage:
      type: object
      properties:
        text: { type: string }
        format: { type: string, enum: [markdown, plain] }
    PluginButton:
      type: object
      properties:
        id: { type: string }
        label: { type: string }
    PluginCard:
      type: object
      properties:
        title: { type: string }
        body: { type: string }
        buttons:
          type: array
          items: { $ref: '#/components/schemas/PluginButton' }
    PluginDefer:
      type: object
      properties:
        job_id: { type: string }
    PluginResponse:
      type: object
      properties:
        messages:
          type: array
          items: { $ref: '#/components/schemas/PluginMessage' }
        cards:
          type: array
          items: { $ref: '#/components/schemas/PluginCard' }
        defer:
          $ref: '#/components/schemas/PluginDefer'